First commit: add tests for MainHandle (successful response, incorrect city, cafe overcount), precode separate to code.go and code_test.go
